requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-js-reverse==0.10.2
environ==1.0
packaging==24.0
pillow==10.3.0
sqlparse==0.5.0
